+++ Characters.cpp
@@ -87,3 +87,57 @@ public:
         }
     }
 };
+
+class Guardian: public CharacterSuper{
+public:
+    Guardian(){
+        health = 100;
+        armor = 75;
+        strength = 100;
+        speed = 25;
+        addToInventory("spear");
+        addToInventory("shield");
+        addSpecialAbility("combat");
+        addSpecialAbility("blocking");
+    }
+};
+
+class Ranger: public CharacterSuper{
+public:
+    Ranger(){
+        health = 100;
+        armor = 50;
+        strength = 25;
+        speed = 100;
+        addToInventory("bow and arrow");
+        addToInventory("knife");
+        addSpecialAbility("stealth");
+        addSpecialAbility("language proficiency");
+    }
+};
+
+class Swordsman: public CharacterSuper{
+public:
+    Swordsman(){
+        health = 100;
+        armor = 100;
+        speed = 50;
+        strength = 50;
+        addToInventory("Heaven Slasher");
+        addToInventory("Manual");
+        addSpecialAbility("Killer Move");
+    }
+};
+
+class Wizard: public CharacterSuper{
+public:
+    Wizard(){
+        health = 100;
+        armor = 25;
+        strength = 50;
+        speed = 75;
+        addToInventory("wand");
+        addToInventory("cloak");
+        addSpecialAbility("fireball");
+    }
+};
