--- Guardian.h
@@ -1,8 +1,0 @@
-#include "Character.h"
-
-class Guardian: public CharacterSuper{
-public:
-    Guardian(){
-
-    }
-};
