--- Ranger.h
@@ -1,8 +1,0 @@
-#include"Character.h" 
-
-class Ranger: public CharacterSuper{
-public:
-    Ranger(){
-
-    }
-};
