--- Swordsman.h
@@ -1,7 +1,0 @@
-#include"Character.h" 
-class Swordsman: public CharacterSuper{
-public:
-    Swordsman(){
-
-    }
-};
