--- Wizard.h
@@ -1,8 +1,0 @@
-#include"Character.h" 
-
-class Wizard: public CharacterSuper{
-public:
-    Wizard(){
-
-    }
-};
